fix: rename top level dir as well
