+++ src/components/AIStateIndicator/AIStateIndicator.tsx
