+++ src/components/AIStateIndicator/hooks/useAIState.ts
