+++ src/components/index.ts
@@ -36,7 +36,7 @@ export * from './UserItem';
 export * from './Window';
 export * from './Threads';
 export * from './ChatView';
-export * from './AITypingIndicatorView';
+export * from './AIStateIndicator';
 
 export { UploadButton } from './ReactFileUtilities';
 export type { UploadButtonProps } from './ReactFileUtilities';
